requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.2.0
djangorestframework==3.12.4
django-rest-knox==4.1.0